+++ src/corewar/g_sub.c
@@ -20,7 +20,7 @@ void				sub(t_cor *cor, t_process *proc)
 	int32_t			value_to_reg; //результат побитового "и", которое нужно записать в регистр (3 арг)
 	int32_t			value1;
 	int32_t			value2;
-	uint8_t			reg3;
+	int32_t			reg3;
 
 	reg3 = cor->map[get_address(proc, get_step(cor, proc, 2), 0)];
 	value1 = get_value(cor, proc, 0);
